+++ packages/shared/src/lib/core/network/classes/evm-explorer-api.class.ts
@@ -2,7 +2,7 @@ import { TokenStandard } from '@core/token/enums'
 import { BaseApi } from '@core/utils/api'
 
 import { DEFAULT_EXPLORER_URLS } from '../constants'
-import { IExplorerApi } from '../interfaces'
+import { IExplorerApi, IExplorerAsset, IExplorerAssetMetadata } from '../interfaces'
 import { NetworkId } from '../types'
 
 export class EvmExplorerApi extends BaseApi implements IExplorerApi {
@@ -25,25 +25,4 @@ export class EvmExplorerApi extends BaseApi implements IExplorerApi {
 
         return response.items.map((asset) => ({ ...asset, type: asset.type.replace('-', '') as TokenStandard }))
     }
-}
-
-export interface IExplorerAsset {
-    token: IExplorerAssetMetadata
-    token_id: string
-    token_instance: unknown
-    value: string
-}
-
-// TODO: Change to real asset metadata and align with other code
-export interface IExplorerAssetMetadata {
-    address: string
-    circulating_market_cap: string
-    decimals: number
-    exchange_rate: string
-    holders: string
-    icon_url: string
-    name: string
-    symbol: string
-    total_supply: string
-    type: TokenStandard
 }
