+++ packages/shared/src/lib/core/network/interfaces/index.ts
@@ -7,6 +7,8 @@ export * from './client-options.interface'
 export * from './connected-chain.interface'
 export * from './evm-addresses.interface'
 export * from './explorer-api.interface'
+export * from './explorer-asset-metadata.interface'
+export * from './explorer-asset.interface'
 export * from './gas-fee-policy.interface'
 export * from './gas-limits.interface'
 export * from './network-status.interface'
